requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15      
aiosignal==1.4.0      
annotated-types==0.7.0
anyio==4.10.0  
attrs==25.3.0
blinker==1.9.0        
cachetools==6.1.0     
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.2.1
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-ollama==0.3.6
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.13
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
ollama==0.5.3
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
qdrant-client==1.15.1
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.47.2
streamlit==1.48.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0